Add BLE DFU transport via WinRT for .NET Framework 4.8
Implements DfuTransportBle using Windows.Devices.Bluetooth, mirroring
noble-nrf-dfu's DfuTransportNoble (GATT discovery, buttonless DFU jump
with rescan/reconnect, PRN/CRC handling) so usage stays consistent
across transports. Also ports several upstream fixes from noble-nrf-dfu
(CRC-mismatch resume-instead-of-abort, per-chunk progress event,
transport Finish()/StartButtonless() hooks, PRN 16-bit bounds check)
into the shared DfuAbstractTransport/DfuTransportPrn/DfuOperation code
used by all transports.

Adds a BLE test console app (test/Nordic.nRF.DFU.BLE.Console) and
documents both in README.md.

+++ src/DfuTransportSerial.cs
@@ -70,7 +70,7 @@ namespace Nordic.nRF.DFU
         private bool _isReady;
         private bool _isGettingReady;
 
-        private DfuTransportSerial(short packetReceiveNotification)
+        private DfuTransportSerial(int packetReceiveNotification)
             : base(packetReceiveNotification)
         {
             _decoder = new SLIPDecoder(this.OnData);
@@ -79,13 +79,13 @@ namespace Nordic.nRF.DFU
             _isGettingReady = false;
         }
 
-        public DfuTransportSerial(SerialPort serialPort, short packetReceiveNotification = 16)
+        public DfuTransportSerial(SerialPort serialPort, int packetReceiveNotification = 16)
             : this(packetReceiveNotification)
         {
             _serialPort = serialPort;
         }
 
-        public DfuTransportSerial(string serialPortName, short packetReceiveNotification = 16)
+        public DfuTransportSerial(string serialPortName, int packetReceiveNotification = 16)
             : this(packetReceiveNotification)
         {
             _serialPortName = serialPortName;
